edna-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
google-cloud-storage==2.14.0
cloud-sql-python-connector==1.5.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
biopython==1.83
python-multipart==0.0.6
pg8000==1.30.3